chore(storybook): update to latest and cleanup
