+++ package.json
@@ -77,8 +77,10 @@
     "@semantic-release/exec": "5.0.0",
     "@semantic-release/git": "9.0.0",
     "@semantic-release/release-notes-generator": "9.0.3",
-    "@storybook/addon-essentials": "6.2.9",
-    "@storybook/react": "6.2.9",
+    "@storybook/addon-actions": "6.3.5",
+    "@storybook/addon-essentials": "6.3.5",
+    "@storybook/addon-links": "6.3.5",
+    "@storybook/react": "6.3.5",
     "@testing-library/jest-dom": "5.14.1",
     "@testing-library/react": "12.0.0",
     "@testing-library/user-event": "13.2.0",
